+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Release Notes
 
 - `TransactionObserver.ignoreDatabaseChangesUntilNextTransaction` allows transaction observers to stop observing the database for the remaining extent of a transaction.
 - GRDB no longer prevents the [truncate optimization](https://www.sqlite.org/lang_delete.html#truncateopt) when no transaction observers are interested in deleted rows.
+- FetchedRecordsController now avoids checking for changes in untracked rowIds.
 - `Row` adopts RandomAccessCollection